Switch from Net::HTTP to HTTPClient (keepalive!)
HTTPClient was chosen primarily because the persistent connections can be used safely across threads. This is a major advantage in for example a rails app where each incoming HTTP connection is handled in a new thread.

This will reduce POST times by 1 round trip even in the case that connections are not reused: Net::HTTP sends post requests in two TCP packets, waiting for the TCP ack before sending the second packet (because TCP_NODELAY is not set) - this is obviously crazy.