Ensure configuration exists
Currently, if important parts of the configuration is missing such as
key, secret or app_id the client will blow up when trying to reference
any of the attributes. An example of this would be executing a `#get`
request and building the proper signature. Since the attributes can be
`nil` we would get an error such as: "Can't convert nil to string".

This commit addresses the problem by ensuring a proper configuration
before building the resource and creating the request.