Authentication for encrypted channels
Authenticated encrypted channels needs to include a step that
generates and packages the shared secret that the client will use
to decrypt messages.

Only in the instance that the channel name has the 'private-encrypted`
prefix will we generate and include a shared_secret key.

If there is no encryption_master_key but the channel is prefixed with
encrypted we will simply send back null.